+++ generators/typescript/sdk/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.0-rc0] - 2024-04-19
+
+- Feature: Bump to v38 of IR and support server-sent events where the events are sent 
+  with a `data: ` prefix and terminated with a new line.  
+
 ## [0.14.1-rc5] - 2024-04-17
 
 - Fix: Code snippets are generated for file upload endpoints using `fs.readStream`. Previously,